+++ MapDriveInator/history.md
@@ -1,5 +1,12 @@
 # History
 
+## v0.0.4
+
+* Added MDI Styles Template map
+* changed obtainBaseFolder logic
+* corrected similarity issue
+* change folder_node recognition logic
+
 ## v0.0.3
 
 * old script version packed as an Add-on